backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
google-generativeai==0.5.4
python-dotenv==1.0.1
websockets==12.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
joblib==1.4.0
